feat: fs exist
* token文件则不创建
* add chalk

+++ src/rule.js
@@ -1,14 +1,18 @@
 const fs = require('fs')
 const path = require('path')
+const chalk = require('chalk')
+const {fsExistsSync} = require('./utils')
 
-const resolve = (dir) => {
-  return path.join(__dirname, '..', dir)
-}
+// const resolve = (dir) => {
+//   return path.join(__dirname, '..', dir)
+// }
 
+// 记录token
 const record = (name, content) => {
+  if (fsExistsSync(name)) return;
   fs.writeFile(name, content, (err, data) => {
     if (err) return console.error(err);
-    console.log(`获取 token 成功！请查看 ${name}`);
+    console.log(chalk.blue(`获取 token 成功！请查看 ${name}`))
     // process.exit(0);
   })
 }
@@ -20,14 +24,14 @@ const records = (request) => {
       record(`token-${(new Date).getMonth()}-${(new Date).getDate()}.txt`, token);
     }
   } catch (error) {
-    // console.log(error)
+    console.log(chalk.red(`获取 token 失败！请查看 ${error}`))
   }
 }
 
 module.exports = {
   // 是否处理https请求
   *beforeDealHttpsRequest(request) {
-    return ['zhudb'].find(host => request.host.indexOf(host) !== -1);
+    return ['zhudb.com'].find(host => request.host.indexOf(host) !== -1);
   },
   *beforeSendRequest(request) {
     records(request);
